+++ Cargo.toml
@@ -78,6 +78,7 @@ encryption = { path = "components/encryption" }
 engine_panic = { path = "components/engine_panic", optional = true }
 engine_rocks = { path = "components/engine_rocks" }
 engine_traits = { path = "components/engine_traits" }
+error_code = { path = "components/error_code" }
 fail = "0.3"
 failure = "0.1"
 fs2 = "0.4"
@@ -230,6 +231,7 @@ members = [
   "components/raftstore",
   "components/into_other",
   "components/encryption",
+  "components/error_code",
 ]
 default-members = ["cmd"]
