+++ components/codec/Cargo.toml
@@ -6,6 +6,7 @@ publish = false
 
 [dependencies]
 byteorder = "1.2"
+error_code = { path = "../error_code" }
 failure = "0.1"
 libc = "0.2"
 static_assertions = { version = "1.0", features = ["nightly"] }
