+++ components/engine_traits/Cargo.toml
@@ -5,6 +5,7 @@ edition = "2018"
 publish = false
 
 [dependencies]
+error_code = { path = "../error_code" }
 hex = "0.4"
 protobuf = "2"
 quick-error = "1.2.3"
