+++ components/error_code/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "error_code"
+version = "0.0.1"
+edition = "2018"
+publish = false
+
+[dependencies]
+raft = { version = "0.6.0-alpha", default-features = false }
+kvproto = { git = "https://github.com/pingcap/kvproto.git", default-features = false }
+
