+++ components/error_code/src/encryption.rs
