+++ components/raftstore/Cargo.toml
@@ -43,6 +43,7 @@ crossbeam = "0.7.2"
 engine_rocks = { path = "../engine_rocks" }
 encryption = { path = "../encryption" }
 engine_traits = { path = "../engine_traits" }
+error_code = { path = "../error_code" }
 fail = "0.3"
 openssl = "0.10"
 fs2 = "0.4"
