+++ components/sst_importer/Cargo.toml
@@ -25,6 +25,7 @@ crc32fast = "1.2"
 encryption = { path = "../encryption" }
 engine_rocks = { path = "../engine_rocks" }
 engine_traits = { path = "../engine_traits" }
+error_code = { path = "../error_code" }
 external_storage = { path = "../external_storage" }
 futures = "0.1"
 futures-cpupool = "0.1"
