+++ components/tidb_query_common/Cargo.toml
@@ -10,6 +10,7 @@ hex = "0.3"
 time = "0.1"
 failure = "0.1"
 derive_more = "0.99.3"
+error_code = { path = "../error_code" }
 tikv_util = { path = "../tikv_util" }
 kvproto = { git = "https://github.com/pingcap/kvproto.git", default-features = false }
 tipb = { git = "https://github.com/pingcap/tipb.git", default-features = false }
