+++ components/tikv_util/Cargo.toml
@@ -18,6 +18,7 @@ configuration = { path = "../configuration" }
 chrono = "0.4"
 crc32fast = "1.2"
 crossbeam = "0.7.2"
+error_code = { path = "../error_code" }
 derive_more = "0.99.3"
 fail = "0.3"
 fs2 = "0.4"
@@ -33,7 +34,6 @@ num_cpus = "1"
 openssl = "0.10"
 prometheus = { version = "0.8", features = ["nightly", "push"] }
 quick-error = "1.2.3"
-raft = { version = "0.6.0-alpha", default-features = false }
 rand = "0.7"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
