+++ src/storage/mvcc/mod.rs
@@ -18,7 +18,7 @@ pub use txn_types::{
 use std::error;
 use std::fmt;
 use std::io;
-use tikv_util::error_code::{self, ErrorCode, ErrorCodeExt};
+use error_code::{self, ErrorCode, ErrorCodeExt};
 use tikv_util::metrics::CRITICAL_ERROR;
 use tikv_util::{panic_when_unexpected_key_or_data, set_panic_mark};
